# openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "San Pablo API",
    "version": "1.0.0",
    "description": "API for art store/bookstore"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/api/imports": {
      "post": {
        "tags": ["Imports"],
        "summary": "Create import job (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV file with product data"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Import job created" },
          "400": { "description": "Invalid file or file missing" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Admin access required" }
        }
      },
      "get": {
        "tags": ["Imports"],
        "summary": "List import jobs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "completed", "failed", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Import jobs list with pagination" },
          "401": { "description": "Authentication required" }
        }
      }
    },
    "/api/imports/{id}": {
      "get": {
        "tags": ["Imports"],
        "summary": "Get import job by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Import job details" },
          "403": { "description": "Access denied" },
          "404": { "description": "Import job not found" }
        }
      }
    },
    "/api/imports/{id}/cancel": {
      "post": {
        "tags": ["Imports"],
        "summary": "Cancel import job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Import job cancelled" },
          "400": { "description": "Job cannot be cancelled" },
          "403": { "description": "Access denied" },
          "404": { "description": "Import job not found" }
        }
      }
    },
...
  }
}